Fix for enqueue_script
code comment says that intent is to put jQuery in head to ensure plugin
compatibility, however the final parameter of wp_register_script or
wp_enqueue_script is $in_footer. When it is set to true, WP will put
scripts in the footer resulting in plugin calls to jQuery before jQuery
has loaded.